PureNote class #20. Implemented but not yet part of the default pack

+++ src/main/java/org/toilelibre/libe/soundtransform/model/library/note/Note.java
@@ -5,15 +5,15 @@ import org.toilelibre.libe.soundtransform.model.exception.SoundTransformExceptio
 
 public interface Note {
 
-    Sound getAttack (int frequency, int channelnum, float lengthInSeconds) throws SoundTransformException;
+    Sound getAttack (float frequency, int channelnum, float lengthInSeconds) throws SoundTransformException;
 
-    Sound getDecay (int frequency, int channelnum, float lengthInSeconds) throws SoundTransformException;
+    Sound getDecay (float frequency, int channelnum, float lengthInSeconds) throws SoundTransformException;
 
     float getFrequency ();
 
     String getName ();
 
-    Sound getRelease (int frequency, int channelnum, float lengthInSeconds) throws SoundTransformException;
+    Sound getRelease (float frequency, int channelnum, float lengthInSeconds) throws SoundTransformException;
 
-    Sound getSustain (int frequency, int channelnum, float lengthInSeconds) throws SoundTransformException;
+    Sound getSustain (float frequency, int channelnum, float lengthInSeconds) throws SoundTransformException;
 }
